requirements.txt
dlib>=19.24.1
fire>=0.5.0
importlib_metadata>=6.0.0
numpy>=1.24.2
openai>=0.27.2
Pillow>=9.5.0
pytest>=7.2.0
requests>=2.28.2
setuptools>=67.2.0
Sphinx>=6.1.3
ultralytics>=8.0.61
